% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSoilSite.R
\name{getSoilSite}
\alias{getSoilSite}
\title{getSoilSite}
\usage{
getSoilSite(lat, lon, global_depth = NULL, set = "phys")
}
\arguments{
\item{lat, }{lon}

\item{global_depth}{(optional), global raster of depth to the R horizon, unfortunately not available anymore in the new update of soilgrids, so, download it separately from https://files.isric.org/soilgrids/former/2017-03-10/data/BDRICM_M_250m_ll.tif}
}
\value{
a numeric vector with percentages of sand (w/w), clay(w/w), SOM(w/w), gravel(v/v); bulk density (g/cm3) and depth (m)
}
\description{
Get soil data from soilgrids REST API, point data is from the 250m resolution dataset, calculates the weigthed mean according layer depths
}
\examples{
getSoilSite(lat=8.7333,lon=-70.883333) 
}
\keyword{soil}
\keyword{texture}
